requirements/common.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/common.in
#
decorator==5.1.1
    # via validators
django-environ==0.9.0
    # via -r requirements/common.in
validators==0.20.0
    # via -r requirements/common.in
